Use proper vim-like arrow keys (hjkl)
The default i3config file used jkl; rather than hjkl to move the focus around
windows or move the windows themselves. This change restores hjkl wherever
movement is involved.

This causes a conflict however. By default, $mod+h is used to change the split
to horizontal. That shortcut is now changed to $mod+b (since b is right next to
v, which is used for the vertical split